att.20

+++ book-service/src/main/java/com/github/flimnk/book_service/integration/exchange/ExchangeDto.java
@@ -5,8 +5,8 @@ import java.math.BigDecimal;
 public record ExchangeDto(
          String from,
          String to,
-         BigDecimal conversionFactor,
-         BigDecimal convertedValue,
+         Double conversionFactor,
+         Double convertedValue,
          String envirolment
 ){
     @Override
